fix: made some last minutes tweaks

+++ src/controllers/areaController.js
@@ -5,7 +5,7 @@ export const getAreas = async (req, res, next) => {
     const { state } = req.query;
     const filter = state ? { state } : {};
     const areas = await Area.find(filter).sort({ name: 1 });
-    res.json(areas);
+    res.json({ areas });
   } catch (error) {
     next(error);
   }
@@ -22,7 +22,7 @@ export const getAreaById = async (req, res, next) => {
       res.status(404);
       throw new Error('Area not found');
     }
-    res.json(area);
+    res.json({ area });
   } catch (error) {
     next(error);
   }
@@ -36,7 +36,7 @@ export const createArea = async (req, res, next) => {
       areaData.location = { type: 'Point', coordinates: location.coordinates };
     }
     const area = await Area.create(areaData);
-    res.status(201).json(area);
+    res.status(201).json({ area });
   } catch (error) {
     next(error);
   }
@@ -64,6 +64,10 @@ export const updateArea = async (req, res, next) => {
 
 export const deleteArea = async (req, res, next) => {
   try {
+    if (!req.params.id.match(/^[0-9a-fA-F]{24}$/)) {
+      res.status(400);
+      throw new Error('Invalid area ID format');
+    }
     const area = await Area.findByIdAndDelete(req.params.id);
     if (!area) {
       res.status(404);
